chore: extract AnimatedCommand to own file
